functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.0.0
google-genai>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pydantic>=2.0.0
flask>=3.0.0
twilio>=8.0.0
